feat(core): add and configure Conventional Changelog
Installs Conventional Changelog to try and keep some sort of method to the
madness that this codebase will inevitably become. We're doing this to try and
keep commit messages tidy, make changelogs feasible to maintain and automate,
and generally communicate where this project is going without significant extra
effort when maintaining.